remove: 불필요한 폴더, 파일 삭제
